Return the state on error
A StateStream is designed for cases where you have some type that
"becomes" a stream for a period of time before reverting back into the
original type. For example, a Postgres connection becomes a stream of
rows before turning back into a connection.

The previous approach was flawed for this use case - we lose access to
the state when an error is encountered. In Postgres for example, it is
possible to encounter an error midway through a set of rows, after which
the connection is back in a useful state ready for more queries.

This causes complications with adaptors that can cause errors to be
generated. The state isn't necessarily ready to be re-exposed at any
arbitrary moment. We could in theory save off the new error and run
through the remainder of the stream, but that is almost certainly
unexpected behavior. For now, I've simply removed all of these methods.